requirements.txt
streamlit
openai-whisper
pydub
scipy
torch
